feat: implement basic graph classes and tests

+++ src/edge/specifics/EdgeSpecificsInterface.php
@@ -71,15 +71,15 @@ interface EdgeSpecificsInterface
      *
      * @param EdgeInterface $edge - the edge
      *
-     * @return double
+     * @return float
      */
-    public function getEdgeWeight(EdgeInterface $edge): double;
+    public function getEdgeWeight(EdgeInterface $edge): float;
     
     /**
      * Set the edge weight
      *
      * @param EdgeInterface $edge - the edge
-     * @param double $weight - the weight
+     * @param float $weight - the weight
      */
-    public function setEdgeWeight(EdgeInterface $edge, double $weight): void;
+    public function setEdgeWeight(EdgeInterface $edge, float $weight): void;
 }
